% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cv_folds}
\alias{cv_folds}
\title{Cross-validation folds}
\usage{
cv_folds(
  target,
  X,
  resample_splits = NULL,
  resample_method = "auto",
  application_created_at = NULL,
  rounded_vintages = TRUE,
  ...
)
}
\description{
Experimental: being redesigned.
}
\keyword{internal}
