% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_features.R
\name{evaluate_features}
\alias{evaluate_features}
\title{Evaluate features}
\usage{
evaluate_features(
  data,
  new_features,
  cv_folds = NA,
  method = NA,
  target = NA,
  ...
)
}
\description{
Experimental: being redesigned.
}
\keyword{internal}
