% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard_training.R
\name{fit_model}
\alias{fit_model}
\title{Fit a model}
\usage{
fit_model(
  rec,
  data,
  method = "ranger",
  ...,
  ml_framework = "caret",
  importance = "impurity"
)
}
\description{
Experimental: being redesigned.
}
\keyword{internal}
