% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woe.R
\name{get_group_stats}
\alias{get_group_stats}
\title{Summarize some statistics}
\usage{
get_group_stats(
  df,
  varname,
  stats = c("woe", "fisher_p_val"),
  binary_outcome = NULL,
  application_status = NULL,
  ...,
  template_mat = NULL,
  score = NULL
)
}
\arguments{
\item{df}{The data.frame object}

\item{varname}{Feature to explore}
}
\value{
table that shows relationship between the feature and the outcome variables
}
\description{
Extract predefined unified variables from the *loan_df* class object
}
\examples{
lalalaaa
}
