% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_aliases.R
\name{map_the_variable}
\alias{map_the_variable}
\title{Maps variable values based on defined aliases and conversion rules}
\usage{
map_the_variable(output_var, out_var_info)
}
\description{
Maps variable values based on defined aliases and conversion rules
}
