% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_features.R
\name{plot_profit_curve}
\alias{plot_profit_curve}
\title{Plot profit curve}
\usage{
plot_profit_curve(
  output,
  worth_good = 100,
  worth_bad = -100,
  variable_set = NA
)
}
\description{
Experimental: being redesigned.
}
\keyword{internal}
