% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_features.R
\name{plot_variable_importance}
\alias{plot_variable_importance}
\title{Plot variable importance}
\usage{
plot_variable_importance(output)
}
\description{
Experimental: being redesigned.
}
\keyword{internal}
