% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loandf.R
\name{loan_df}
\alias{loan_df}
\title{Create a data frame object easily recognized by the other `loan` package functions with the relevant credit-loan-portfolio-indicative.}
\usage{
loan_df(
  x,
  client_id = NA,
  application_id = NA,
  loan_id = NA,
  application_created_at = NA,
  application_status = NA,
  loan_status = NA,
  target = NA,
  data_downloaded_at = Sys.time()
)
}
\arguments{
\item{x}{A data frame or matrix of the credit loan/application portfolio data set where each row indicates individual loans or applications but the columns -
all the information related to the corresponding loan/application including, but not limited to, observations data (characteristics available before taking a loan issuance decision),
client/application/loan id fields, created at or registered at time stamp fields, current days passed due, and more.}

\item{client_id}{The column name that contains unique client identification number. Leave it NA if not available.}

\item{application_id}{The column name that contains unique loan application identification number. Leave it NA if not available.}

\item{loan_id}{The column name that contains unique loan identification number. Leave it NA if not available.}

\item{application_created_at}{The column name that indicates the application's creation datetime. For a higher accuracy, it is recommended
that the column is `POSIXct` class but `Date` class will also work. Leave it NA if not available.}

\item{application_status}{The column name that indicates the loans applications' statuses. The column must be a character or factor class. The recognized
application statuses: \n
 * `ISSUED`: The application has been accepted resulted in an issued loan. \n
 * `REJECTED`: The application has been rejected due to the internal lending rules and/or the scorecard. \n
 * `NOT_TAKEN_UP`: The client application has been approved by the lending processes but the client hasn't approved herself the final offer resulting in the application cancellation. \n
 * `WAITING_FOR_APPROVAL`: Similarly as `NOT_TAKEN_UP` the client has been approved by the lending rules but the loan hasn't been issued yet because the client hasn't accepted the
 final offer. The difference is that `WAITING_FOR_APPROVAL` applications still can be approved by the client resulting in `ISSUED` (if the client accepts) status or
 `NOT_TAKEN_UP` (client rejects the final offer or the offer expires) status. \n
 * `CANCELLED`: The client has started creating the application but it was cancelled (most often, unfinished; therefore, not submitted) even before the evaluation. \n
 * `IN_PROCESS`: The application currently is being evaluated and, in the near future, the status will change to the `REJECTED` (if it did no pass the internal evaluation),
 `WAITING_FOR_APPROVAL` (if it passed the internal evaluation and the offer needs to be accepted by the client), or `ISSUED` (if it passed the internal evaluation and additional
 offer acceptance by the client is not required).}

\item{data_downloaded_at}{The data set acquisition date. May be used for the forecasting related analysis where it becomes relevant. The default is the current date.}
}
\value{
A `loan_df` object, which can be easily used with the other `loan` package functions. A `loan_df`object is also a data frame.
}
\description{
The function takes an existing data frame containing columns that are typically available for the loan portfolio data sets
(e.g., loan status, application created_at, client id, first payment due, etc.) and turns them into a new data frame object recognized by the other
`loan` package functions for the automatic and easy data analysis.
}
\examples{
Lalallaaaa
a <- 1:100
}
